PullRequest: graalpython/2119

+++ mx.graalpython/mx_graalpython.py
@@ -140,7 +140,7 @@ def python(args, **kwargs):
     do_run_python(args, **kwargs)
 
 
-def do_run_python(args, extra_vm_args=None, env=None, jdk=None, extra_dists=None, cp_prefix=None, cp_suffix=None, main_class=GRAALPYTHON_MAIN_CLASS, **kwargs):
+def do_run_python(args, extra_vm_args=None, env=None, jdk=None, extra_dists=None, cp_prefix=None, cp_suffix=None, main_class=GRAALPYTHON_MAIN_CLASS, minimal=False, **kwargs):
     experimental_opt_added = False
     if not any(arg.startswith("--python.CAPI") for arg in args):
         capi_home = _get_capi_home()
@@ -164,9 +164,16 @@ def do_run_python(args, extra_vm_args=None, env=None, jdk=None, extra_dists=None
         elif check_vm_env == '0':
             check_vm()
 
-    dists = ['GRAALPYTHON', 'TRUFFLE_NFI', 'SULONG_NATIVE']
+    if minimal:
+        x = [x for x in SUITE.dists if x.name == "GRAALPYTHON"][0]
+        dists = [dep for dep in x.deps if dep.isJavaProject() or dep.isJARDistribution()]
+    else:
+        dists = ['GRAALPYTHON']
+    dists += ['TRUFFLE_NFI', 'SULONG_NATIVE']
 
     vm_args, graalpython_args = mx.extract_VM_args(args, useDoubleDash=True, defaultAllVMArgs=False)
+    if minimal:
+        vm_args.insert(0, f"-Dorg.graalvm.language.python.home={_dev_pythonhome()}")
     graalpython_args, additional_dists = _extract_graalpython_internal_options(graalpython_args)
     dists += additional_dists
 
@@ -1048,8 +1055,17 @@ def _get_src_dir(projectname):
     mx.abort("Could not find src dir for project %s" % projectname)
 
 
+def _get_output_root(projectname):
+    for suite in mx.suites():
+        for p in suite.projects:
+            if p.name == projectname:
+                return p.get_output_root()
+    mx.abort("Could not find out dir for project %s" % projectname)
+
+
 mx_subst.path_substitutions.register_with_arg('suite', _get_suite_dir)
 mx_subst.path_substitutions.register_with_arg('src_dir', _get_src_dir)
+mx_subst.path_substitutions.register_with_arg('output_root', _get_output_root)
 
 
 def delete_self_if_testdownstream(args):
@@ -1858,7 +1874,7 @@ def python_build_watch(args):
             mx.log("Build done.")
 
 
-class GraalpythonCAPIBuildTask(mx.ProjectBuildTask):
+class GraalpythonBuildTask(mx.ProjectBuildTask):
     class PrefixingOutput():
         def __init__(self, prefix, printfunc):
             self.prefix = "[" + prefix + "] "
@@ -1870,61 +1886,44 @@ class GraalpythonCAPIBuildTask(mx.ProjectBuildTask):
 
     def __init__(self, args, project):
         jobs = min(mx.cpu_count(), 8)
-        super(GraalpythonCAPIBuildTask, self).__init__(args, jobs, project)
+        super().__init__(args, jobs, project)
 
     def __str__(self):
-        return 'Building C API project {} with setuptools'.format(self.subject.name)
-
-    def run(self, args, env=None, cwd=None, **kwargs):
-        env = env.copy() if env else os.environ.copy()
-        # n.b.: we don't want derived projects to also have to depend on our build env vars
-        env.update(mx.dependency("com.oracle.graal.python.cext").getBuildEnv())
-        env.update(self.subject.getBuildEnv())
-
-        # distutils will honor env variables CC, CFLAGS, LDFLAGS but we won't allow to change them
-        for var in ["CC", "CFLAGS", "LDFLAGS"]:
-            env.pop(var, None)
-        args.insert(0, '--PosixModuleBackend=java')
-        return do_run_python(args, env=env, cwd=cwd, out=self.PrefixingOutput(self.subject.name, mx.log), err=self.PrefixingOutput(self.subject.name, mx.log_error), **kwargs)
-
-    def _dev_headers_dir(self):
-        return os.path.join(SUITE.dir, "graalpython", "include")
-
-    def _prepare_headers(self):
-        target_dir = self._dev_headers_dir()
-        if os.path.exists(target_dir):
-            shutil.rmtree(target_dir)
-        mx.logv("Preparing header files (dest: {!s})".format(target_dir))
-        shutil.copytree(os.path.join(self.src_dir(), "include"), target_dir)
-        shutil.copy(os.path.join(mx.dependency("SULONG_LEGACY").get_output(), "include", "truffle.h"), target_dir)
+        return 'Building project {}'.format(self.subject.name)
 
     def build(self):
-        self._prepare_headers()
+        if not self.args.force and not self.args.all and not self.needsBuild(None)[0]:
+            mx.log(f"Refusing build of {self.subject.name}. Use e.g. `mx -f --only {self.subject.name}' to force a build.")
+            return True
+        args = [mx_subst.path_substitutions.substitute(a, dependency=self) for a in self.subject.args]
+        return self.run(args)
 
-        # n.b.: we do the following to ensure that there's a directory when the
-        # importlib PathFinder initializes it's directory finders
-        mx.ensure_dir_exists(os.path.join(self.subject.get_output_root(), "modules"))
-
-        cwd = os.path.join(self.subject.get_output_root(), "mxbuild_temp")
+    def run(self, args, env=None, cwd=None, **kwargs):
+        cwd = cwd or os.path.join(self.subject.get_output_root(), "mxbuild_temp")
 
         if os.path.exists("/dev/null"):
             pycache_dir = "/dev/null"
         else:
             pycache_dir = os.path.join(self.subject.get_output_root(), "__pycache__")
 
-        args = []
         if mx._opts.verbose:
-            args.append("-v")
+            args.insert(0, "-v")
         else:
             # always add "-q" if not verbose to suppress hello message
-            args.append("-q")
+            args.insert(0, "-q")
 
-        args += ["--python.PyCachePrefix=" + pycache_dir,
-                 "-B",
-                 "-S", os.path.join(self.src_dir(), "setup.py"),
-                 self.subject.get_output_root()]
+        args[:0] = [
+            f"--python.PyCachePrefix={pycache_dir}",
+            "-B",
+            "-S"
+        ]
         mx.ensure_dir_exists(cwd)
-        rc = self.run(args, cwd=cwd)
+
+        env = env.copy() if env else os.environ.copy()
+        env.update(self.subject.getBuildEnv())
+        args.insert(0, '--PosixModuleBackend=java')
+        rc = do_run_python(args, env=env, cwd=cwd, minimal=True, out=self.PrefixingOutput(self.subject.name, mx.log), err=self.PrefixingOutput(self.subject.name, mx.log_error), **kwargs)
+
         shutil.rmtree(cwd) # remove the temporary build files
         # if we're just running style tests, this is allowed to fail
         if os.environ.get("BUILD_NAME") == "python-style":
@@ -1954,7 +1953,7 @@ class GraalpythonCAPIBuildTask(mx.ProjectBuildTask):
         if tsOldest == sys.maxsize:
             tsOldest = 0
         if tsOldest < tsNewest:
-            self.clean() # we clean here, because setuptools doesn't check timestamps
+            self.clean(forBuild="reallyForBuild") # we clean here, because setuptools doesn't check timestamps
             if newestFile and oldestFile:
                 return (True, "rebuild needed, %s newer than %s" % (newestFile, oldestFile))
             else:
@@ -1966,28 +1965,77 @@ class GraalpythonCAPIBuildTask(mx.ProjectBuildTask):
         return None
 
     def clean(self, forBuild=False):
+        if forBuild == "reallyForBuild":
+            try:
+                shutil.rmtree(self.subject.get_output_root())
+            except BaseException:
+                return 1
+        return 0
+
+
+class GraalpythonCAPIBuildTask(GraalpythonBuildTask):
+    def run(self, args, env=None, cwd=None, **kwargs):
+        env = env.copy() if env else os.environ.copy()
+        # n.b.: we don't want derived projects to also have to depend on our build env vars
+        env.update(mx.dependency("com.oracle.graal.python.cext").getBuildEnv())
+        env.update(self.subject.getBuildEnv())
+
+        # distutils will honor env variables CC, CFLAGS, LDFLAGS but we won't allow to change them
+        for var in ["CC", "CFLAGS", "LDFLAGS"]:
+            env.pop(var, None)
+        return super().run(args, env=env, cwd=cwd, **kwargs)
+
+    def _dev_headers_dir(self):
+        return os.path.join(SUITE.dir, "graalpython", "include")
+
+    def _prepare_headers(self):
+        target_dir = self._dev_headers_dir()
+        if os.path.exists(target_dir):
+            shutil.rmtree(target_dir)
+        mx.logv("Preparing header files (dest: {!s})".format(target_dir))
+        shutil.copytree(os.path.join(self.src_dir(), "include"), target_dir)
+        shutil.copy(os.path.join(mx.dependency("SULONG_LEGACY").get_output(), "include", "truffle.h"), target_dir)
+
+    def build(self):
+        self._prepare_headers()
+        # n.b.: we do the following to ensure that there's a directory when the
+        # importlib PathFinder initializes it's directory finders
+        mx.ensure_dir_exists(os.path.join(self.subject.get_output_root(), "modules"))
+        # TODO: backwards compat, remove once EE is updated
+        if not hasattr(self.subject, "args"):
+            self.subject.args = [
+                os.path.join(self.src_dir(), "setup.py"),
+                self.subject.get_output_root()
+            ]
+        return super().build()
+
+    def clean(self, forBuild=False):
         result = 0
-        try:
-            shutil.rmtree(self._dev_headers_dir())
-        except BaseException:
-            result = 1
-        try:
-            shutil.rmtree(self.subject.get_output_root())
-        except BaseException:
-            result = 1
-        return result
+        if not forBuild:
+            try:
+                shutil.rmtree(self._dev_headers_dir())
+            except BaseException:
+                result = 1
+        return max(result, super().clean(forBuild=forBuild))
 
 
-class GraalpythonCAPIProject(mx.Project):
-    def __init__(self, suite, name, subDir, srcDirs, deps, workingSets, d, theLicense=None, **kwargs):
+class GraalpythonProject(mx.ArchivableProject):
+    def __init__(self, suite, name, subDir, srcDirs, deps, workingSets, d, theLicense=None, **kwargs): # pylint: disable=super-init-not-called
         context = 'project ' + name
         self.buildDependencies = mx.Suite._pop_list(kwargs, 'buildDependencies', context)
-        if mx.suite("sulong-managed", fatalIfMissing=False) is not None:
-            self.buildDependencies.append('sulong-managed:SULONG_MANAGED_HOME')
-        super(GraalpythonCAPIProject, self).__init__(suite, name, subDir, srcDirs, deps, workingSets, d, theLicense, **kwargs)
+        mx.Project.__init__(self, suite, name, subDir, srcDirs, deps, workingSets, d, theLicense, **kwargs)
 
     def getOutput(self, replaceVar=mx_subst.results_substitutions):
         return self.get_output_root()
+
+    def output_dir(self):
+        return self.getOutput()
+
+    def archive_prefix(self):
+        return ""
+
+    def getResults(self):
+        return [x[0] for x in self.getArchivableResults(use_relpath=False)]
 
     def getArchivableResults(self, use_relpath=True, single=False):
         if single:
@@ -2002,7 +2050,7 @@ class GraalpythonCAPIProject(mx.Project):
                     yield fullname, name
 
     def getBuildTask(self, args):
-        return GraalpythonCAPIBuildTask(args, self)
+        return GraalpythonBuildTask(args, self)
 
     def getBuildEnv(self, replaceVar=mx_subst.path_substitutions):
         ret = {}
@@ -2010,6 +2058,16 @@ class GraalpythonCAPIProject(mx.Project):
             for key, value in self.buildEnv.items():
                 ret[key] = replaceVar.substitute(value, dependency=self)
         return ret
+
+
+class GraalpythonCAPIProject(GraalpythonProject):
+    def __init__(self, suite, name, subDir, srcDirs, deps, workingSets, d, theLicense=None, **kwargs):
+        super().__init__(suite, name, subDir, srcDirs, deps, workingSets, d, theLicense, **kwargs)
+        if mx.suite("sulong-managed", fatalIfMissing=False) is not None:
+            self.buildDependencies.append('sulong-managed:SULONG_MANAGED_HOME')
+
+    def getBuildTask(self, args):
+        return GraalpythonCAPIBuildTask(args, self)
 
 
 def checkout_find_version_for_graalvm(args):
